Predictive-Transaction-Intelligence-using-for-BFSI-backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
transformers
torch
python-dotenv
pydantic
joblib
pyyaml
seaborn
matplotlib
python-dotenv
